Refactor test structure
It’s common to add a test helper file so that we can put any
configuration for our tests in one place, without having to repeat it.

The main benefit here is that we only need to require one file in each
of our tests, and that file will require ‘rspec’ and ‘pry’ for us,
along with all the models that we need.

Can you refactor the spec_helper further so that it automatically
requires any file in the lib directory without having to list them all?